Created components & changed layout to feature a category page
