+++ screens/BodyParts.js
@@ -1,0 +1,12 @@
+import { View, Text } from "react-native";
+import React from "react";
+
+const BodyParts = () => {
+  return (
+    <View>
+      <Text>BodyParts</Text>
+    </View>
+  );
+};
+
+export default BodyParts;
